+++ repositoryminer-metrics/src/main/java/org/repositoryminer/metrics/codesmell/BrainClass.java
@@ -14,7 +14,7 @@ public class BrainClass extends CodeSmell {
 	private static final CodeSmellId[] REQUIRED_CODESMELLS = { CodeSmellId.BRAIN_METHOD };
 
 	private int wmcThreshold = 47;
-	private float tccThreshold = 0.5f;
+	private double tccThreshold = 0.5;
 	private int brainMethodThreshold = 1;
 	private int locThreshold = 195;
 
@@ -22,7 +22,7 @@ public class BrainClass extends CodeSmell {
 		init();
 	}
 
-	public BrainClass(int wmcThreshold, float tccThreshold, int brainMethodThreshold, int locThreshold) {
+	public BrainClass(int wmcThreshold, double tccThreshold, int brainMethodThreshold, int locThreshold) {
 		this.wmcThreshold = wmcThreshold;
 		this.tccThreshold = tccThreshold;
 		this.brainMethodThreshold = brainMethodThreshold;
@@ -46,7 +46,7 @@ public class BrainClass extends CodeSmell {
 	public void detect(FileReport fileReport, ProjectReport projectReport) {
 		for (ClassReport cr : fileReport.getClasses()) {
 			int wmc = cr.getMetricsReport().getCodeMetric(CodeMetricId.WMC, Integer.class);
-			float tcc = cr.getMetricsReport().getCodeMetric(CodeMetricId.TCC, Float.class);
+			double tcc = cr.getMetricsReport().getCodeMetric(CodeMetricId.TCC, Double.class);
 			int loc = cr.getMetricsReport().getCodeMetric(CodeMetricId.LOC, Integer.class);
 			int nbm = 0;
 
@@ -62,7 +62,7 @@ public class BrainClass extends CodeSmell {
 		}
 	}
 
-	public boolean detect(int nbm, int loc, int wmc, float tcc) {
+	public boolean detect(int nbm, int loc, int wmc, double tcc) {
 		boolean exp1 = nbm > brainMethodThreshold && loc >= locThreshold;
 		boolean exp2 = nbm == brainMethodThreshold && loc >= (2 * locThreshold) && wmc >= (2 * wmcThreshold);
 		boolean exp3 = wmc >= wmcThreshold && tcc < tccThreshold;
@@ -79,11 +79,11 @@ public class BrainClass extends CodeSmell {
 		this.wmcThreshold = wmcThreshold;
 	}
 
-	public float getTccThreshold() {
+	public double getTccThreshold() {
 		return tccThreshold;
 	}
 
-	public void setTccThreshold(float tccThreshold) {
+	public void setTccThreshold(double tccThreshold) {
 		this.tccThreshold = tccThreshold;
 	}
